Move spatialhash to _core
