+++ parcels/_core/spatialhash.py
